feat: rename custom annotation to shape annotation
